requirements.txt
#PYTHON
pandas
scikit-learn
python-dotenv
ipykernel

# SQL 
mysql-connector-python

#YAML
pyyaml